Add files via upload
This data file is supplementary to the file named "DATA_TABLE_WITHOUT_THRUST.xlsx" as they contain the same dynamics model and same testing conditions. The difference with this file, however, is that it now includes incremental corrections to the trajectory via a chemical propulsion system. This system simply approximates a flight control system that provides a fixed delta-V in fixed time intervals. The max delta-V is 10 m/s (but scales down for smaller errors) and this correction occurs instantaneously every 15 seconds. The maximum number of impulsive correction maneuvers is 25. It is expected that the size of the landing ellipses will decrease due to the control system.